Message Doesn't Support 64-bit Thread IDs #1913

+++ Foundation/include/Poco/Message.h
@@ -193,7 +193,7 @@ private:
 	std::string _text;
 	Priority    _prio;
 	Timestamp   _time;
-	int         _tid;
+	long        _tid;
 	std::string _thread;
 	long        _pid;
 	const char* _file;
